requirements.txt
# FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2==2.9.10
alembic==1.14.1
greenlet==3.1.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Celery
celery[redis]==5.4.0
redis==5.2.1

# HTTP client
httpx==0.28.1

# DNS
dnspython==2.7.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx  # also used as test client

# Utils
pydantic[email-validation]==2.10.4
email-validator==2.2.0
pydantic-settings==2.7.1
python-dotenv==1.0.1
